+++ src/utils.ts
@@ -18,6 +18,16 @@ export function delay(ms: number) {
   return new Promise<void>(r => setTimeout(r, ms));
 }
 
+export function block<T = void>() {
+  let resolve: (v: T) => void;
+  let reject: (e: any) => void;
+  let promise = new Promise<T>((res, rej) => {
+    resolve = res;
+    reject = rej;
+  });
+  return { promise, resolve: resolve!, reject: reject! };
+}
+
 export interface ExternalPluginOptions {
   /**
    * Passed to `onResolve()`, mark them as external.